Abstracting the SQL connectors

+++ server/prisma-rs/query-engine/core/src/builders/mod.rs
@@ -16,13 +16,13 @@ pub use single::*;
 
 use self::inflector::Inflector;
 use crate::{CoreError, CoreResult, ReadQuery};
-use ::inflector::Inflector as RustInflector;
 use connector::{filter::NodeSelector, QueryArguments};
 use graphql_parser::query::{Field, Selection, Value};
 use prisma_models::{
     Field as ModelField, GraphqlId, ModelRef, OrderBy, PrismaValue, RelationFieldRef, SchemaRef, SelectedField,
     SelectedFields, SelectedRelationField, SelectedScalarField, SortOrder,
 };
+use rust_inflector::Inflector as RustInflector;
 
 use std::{collections::BTreeMap, sync::Arc};
 use uuid::Uuid;
@@ -129,7 +129,7 @@ pub trait BuilderExt {
                 Ok(val) => PrismaValue::Json(val),
                 _ => PrismaValue::String(s.clone()),
             },
-            Value::Int(i) => PrismaValue::Int(i.as_i64().unwrap() as i32),
+            Value::Int(i) => PrismaValue::Int(i.as_i64().unwrap()),
             _ => unimplemented!(),
         }
     }
